Refactoring work and initial HLS/HDS support
Large commit. Sorry for that.

We can now know that there are other things than RTMP, and identify
and extract such URLs. Currently, we can't do much more though. HLS
may be playable with mplayer, but not sure how mplayer selects
quality. I usually do the following:

 $ svtplay <svtplay url>
 (yank the url)
 $ curl -s <the yanked url>
 (look through the M3U file after the quality i want)

My plan is to implement HLS as a separate module, and then integrate
it here. HDS is an other story.